+++ backend/src/middleware/GenerateSchedule/tempCodeRunnerFile.ts
@@ -1,0 +1,5 @@
+await getSchedule(activityDistance)
+  //   .then((result) => {
+  //     schedule = result;
+  //   })
+  //   .catch((error) => console.error("Error creating schedule:", error));
